Update build image to use openjdk base instead of fedora
Altered the build image Dockerfile to use openjdk:8-jdk instead of fedora. This changes the base OS to Debian. Either base actually works, so saved the old Dockerfile into a new folder named fedora_java_maven in case someone prefers to use that.

Attempted to use Alpine as the build base also, using the openjdk alpine base. It works for most of the build but fails to build the UI because Maven pulls in a nodejs binary incompatible with Alpine. ( it is dependent on glibc, and Alpine uses musl ) It is possible to workaround this issue by adding in musl on top of Alpine, but it was too much hassle and I didn't pursue further. It isn't tremendously important as the build process already pulls in so many Maven modules that their size dwarfs the base os layer size.